Fixing TODOs
Removed non-relevant TODOs
Fixed code duplication
