+++ backend/dataall/api/Objects/ShareObject/schema.py
@@ -1,4 +1,5 @@
 from .resolvers import *
+from dataall.api.Objects.Environment.resolvers import resolve_environment
 
 ShareableObject = gql.Union(
     name='ShareableObject',
